added message types to client sending

+++ Server/src/main/java/com/poo/musicbroadcaster/model/client/SeekMessage.java
@@ -1,9 +1,9 @@
 package com.poo.musicbroadcaster.model.client;
 
 public class SeekMessage {
-	private int milliseconds;
+	private long milliseconds;
 	
-	public int getMilliseconds() {
+	public long getMilliseconds() {
 		return this.milliseconds;
 	}
 }
